services/inference-api/requirements.txt
setuptools>=70.0.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
redis==5.0.8
prometheus-client==0.20.0

opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0

pydantic>=2.7.0
httpx>=0.27.0
pytest>=8.2.0

--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.10.0+cpu
transformers==4.44.0
